refactor(world): move IUltimaMapProvider into Server.Abstractions/Server
Moongate.Server needs to resolve a Map from a MapType for the upcoming
movement tile-query service, but that capability lived in Http.Plugin.
Moving the interface to Server.Abstractions (already referenced by
Http.Plugin) and the implementation to Server avoids a circular project
reference, since Http.Plugin already references Server.

- Create IUltimaMapProvider in Moongate.Server.Abstractions.Interfaces.World
- Move UltimaMapProvider to Moongate.Server.Services.World
- Register in Server's composition root (Program.cs)
- Add explicit Moongate.Ultima project reference to Moongate.Server.csproj
- Promote StubMapProvider test double for reuse in MapTileServiceTests
- Update all consumers (Http.Plugin services/endpoints, tests)
