+++ src/Moongate.Http.Plugin/Endpoints/Maps/MapImageEndpoints.cs
@@ -5,6 +5,7 @@ using Moongate.Http.Plugin.Data;
 using Moongate.Http.Plugin.Interfaces.Endpoints;
 using Moongate.Http.Plugin.Interfaces.Maps;
 using Moongate.Http.Plugin.Types;
+using Moongate.Server.Abstractions.Interfaces.World;
 using Moongate.UO.Data.Types;
 
 namespace Moongate.Http.Plugin.Endpoints.Maps;
