+++ src/Moongate.Http.Plugin/Services/Maps/MapImageService.cs
@@ -2,6 +2,7 @@ using Moongate.Http.Plugin.Data;
 using Moongate.Http.Plugin.Interfaces.Maps;
 using Moongate.Http.Plugin.Interfaces.Ultima;
 using Moongate.Http.Plugin.Types;
+using Moongate.Server.Abstractions.Interfaces.World;
 using Moongate.Ultima.Imaging;
 using Moongate.Ultima.Tiles;
 using Moongate.Ultima.Types;
