+++ src/Moongate.Server.Abstractions/Interfaces/World/IUltimaMapProvider.cs
@@ -1,7 +1,7 @@
 using Moongate.Ultima.Maps;
 using Moongate.UO.Data.Types;
 
-namespace Moongate.Http.Plugin.Interfaces.Maps;
+namespace Moongate.Server.Abstractions.Interfaces.World;
 
 /// <summary>
 /// The facets the shard's client files describe. Exists as an interface because Ultima's facets are static
