+++ src/Moongate.Server/Services/World/UltimaMapProvider.cs
@@ -1,13 +1,13 @@
-using Moongate.Http.Plugin.Interfaces.Maps;
+using Moongate.Server.Abstractions.Interfaces.World;
 using Moongate.Ultima.Maps;
 using Moongate.UO.Data.Types;
 
-namespace Moongate.Http.Plugin.Services.Maps;
+namespace Moongate.Server.Services.World;
 
 /// <summary>
 /// Ultima's own facets, by <see cref="MapType" />. The sizes are Ultima's, not MapDefinitions': Ultima
-/// hardcodes Felucca at 6144×4096 while MapDefinitions says 7168, and the renderer can only draw what
-/// Ultima's Map object describes. On a shard with post-ML client files this renders the older extent.
+/// hardcodes Felucca at 6144×4096 while MapDefinitions says 7168, and callers can only query what
+/// Ultima's Map object describes. On a shard with post-ML client files this reports the older extent.
 /// </summary>
 public sealed class UltimaMapProvider : IUltimaMapProvider
 {
@@ -24,8 +24,8 @@ public sealed class UltimaMapProvider : IUltimaMapProvider
     public Map? Get(MapType facet)
         => facet switch
         {
-            MapType.Felucca => Map.Felucca,
-            MapType.Trammel => Map.Trammel,
+            MapType.Felucca  => Map.Felucca,
+            MapType.Trammel  => Map.Trammel,
             MapType.Ilshenar => Map.Ilshenar,
             MapType.Malas    => Map.Malas,
             MapType.Tokuno   => Map.Tokuno,
